7.10. Отрисуй меня полностью (часть 1)
